menu "Aliyun iotkit embedded" change int16_t to float change esp_tls_conn_new esp_tls_conn_new_sync change scan_config={0}

+++ examples/solutions/smart_light/components/lightbulb/lightbulb.c
@@ -78,7 +78,7 @@ uint32_t duties[PWM_IO_NUM] = {
 };
 
 // phase table, (phase/180)*depth
-int16_t phase[PWM_IO_NUM] = {
+float phase[PWM_IO_NUM] = {
     0, 0, 50,
 };
